Worked on repositories

+++ Cryptocop.Models/Entities/Order.cs
@@ -29,6 +29,6 @@ public class Order
 
     public DateTime OrderDate { get; set; }
 
-    public decimal TotalPrice { get; set; }
+    public float TotalPrice { get; set; }
 
 }
