+++ Cryptocop.Models/Entities/OrderItem.cs
@@ -17,9 +17,9 @@ public class OrderItem
 
     public string ProductIdentifier { get; set;}
 
-    public decimal Quantity { get; set; }
+    public float Quantity { get; set; }
 
-    public decimal UnitPrice { get; set; }
+    public float UnitPrice { get; set; }
 
-    public decimal TotalPrice { get; set; }
+    public float TotalPrice { get; set; }
 }
